move diff/changelog specific code to diff.go
